pyproject.toml
[project]
name = "urbanpulse-iot"
version = "1.0.0"
description = "Extensible urban IoT monitoring platform"
requires-python = ">=3.11"

dependencies = [
    "paho-mqtt>=2.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.42.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "redis>=5.0.0",
    "prometheus-client>=0.25.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "aiosqlite>=0.22.1",
    "ruff>=0.15.6",
    "isort>=8.0.1",
    "pyupgrade>=3.21.2",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.pytest.ini_options]
addopts = ["-ra"]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
